+++ packages/akoya-adapter/playwright/akoya.playwright.ts
@@ -42,8 +42,10 @@ test("connects to mikomo bank with oAuth", async ({ page }) => {
       const obj = (await msg.args()[0].jsonValue())?.message;
       if (obj?.type === "connect/memberConnected") {
         clearTimeout(timer);
-        expect(obj.metadata.user_guid).toEqual(userId);
-        expect(obj.metadata.member_guid).toEqual("mikomo");
+        expect(obj.metadata.user_guid).toBeUndefined();
+        expect(obj.metadata.aggregatorUserId).toEqual(userId);
+        expect(obj.metadata.member_guid).toBeUndefined();
+        expect(obj.metadata.connectionId).toEqual("mikomo");
         expect(obj.metadata.aggregator).toEqual("akoya_sandbox");
         expect(obj.metadata.akoyaAuthCode).not.toBeNull();